:recycle: Integarte UserInterface to AppState
