+++ lib/model/user.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:Pilll/model/pill_sheet.dart';
 import 'package:Pilll/model/setting.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:meta/meta.dart';
 
@@ -21,7 +20,6 @@ extension UserFirestoreFieldKeys on String {
 
 class User {
   static final path = "users";
-  static User _cache;
 
   final String anonymousUserID;
   String get documentID => anonymousUserID;
@@ -34,7 +32,7 @@ class User {
     @required this.currentPillSheet,
   });
 
-  static User _map(Map<String, dynamic> firestoreDocumentData) {
+  static User map(Map<String, dynamic> firestoreDocumentData) {
     return User._(
       anonymousUserID:
           firestoreDocumentData[UserFirestoreFieldKeys.anonymouseUserID],
@@ -53,46 +51,8 @@ class User {
     );
   }
 
-  static User user() {
-    if (_cache == null) throw UserNotFound();
-    return _cache;
-  }
-
-  static Future<User> fetch() {
-    if (_cache != null) {
-      return Future.value(_cache);
-    }
-    return FirebaseFirestore.instance
-        .collection(User.path)
-        .doc(FirebaseAuth.instance.currentUser.uid)
-        .get()
-        .then((document) {
-      if (!document.exists) {
-        throw UserNotFound();
-      }
-      var user = User._map(document.data());
-      assert(_cache == null);
-      _cache = user;
-      return user;
-    });
-  }
-
   DocumentReference documentReference() {
     return FirebaseFirestore.instance.collection(User.path).doc(documentID);
-  }
-
-  static Future<User> create() {
-    return FirebaseFirestore.instance
-        .collection(User.path)
-        .doc(FirebaseAuth.instance.currentUser.uid)
-        .set(
-      {
-        UserFirestoreFieldKeys.anonymouseUserID:
-            FirebaseAuth.instance.currentUser.uid,
-      },
-    ).then((_) {
-      return User.fetch();
-    });
   }
 
   Future<void> deleteCurrentPillSheet() {
@@ -102,17 +62,5 @@ class User {
       },
       SetOptions(merge: true),
     ).then((_) => this.currentPillSheet = null);
-  }
-}
-
-extension UserInterface on User {
-  static Future<User> fetchOrCreateUser() {
-    return User.fetch().catchError((error) {
-      if (error is UserNotFound) {
-        return User.create();
-      }
-      throw FormatException(
-          "cause exception when failed fetch and create user for $error");
-    });
   }
 }
